Directory pattern bracketed expressions (#3238)
