+++ src/main/java/org/jabref/gui/preftabs/ImportSettingsTab.java
@@ -23,8 +23,8 @@ import com.jgoodies.forms.layout.FormLayout;
 
 public class ImportSettingsTab extends JPanel implements PrefsTab {
 
-    public static final String[] DEFAULT_FILENAMEPATTERNS = new String[] {"\\bibtexkey",
-            "\\bibtexkey\\begin{title} - \\format[RemoveBrackets]{\\title}\\end{title}"};
+    public static final String[] DEFAULT_FILENAMEPATTERNS = new String[] {"[bibtexkey]",
+            "[bibtexkey] - [title]"};
 
     public static final int DEFAULT_STYLE = ImportDialog.CONTENT;
     private static final String[] DEFAULT_FILENAMEPATTERNS_DISPLAY = new String[] {"bibtexkey", "bibtexkey - title",};
