+++ src/test/java/org/jabref/logic/util/io/RegExpBasedFileFinderTests.java
@@ -64,6 +64,44 @@ public class RegExpBasedFileFinderTests {
     }
 
     @Test
+    public void testYearAuthFirspageFindFiles() {
+        //given
+        List<String> extensions = Collections.singletonList("pdf");
+
+        List<Path> dirs = Collections.singletonList(Paths.get(filesDirectory));
+        RegExpBasedFileFinder fileFinder = new RegExpBasedFileFinder("**/[year]_[auth]_[firstpage].*\\\\.[extension]", ',');
+
+        //when
+        List<Path> result = fileFinder.findAssociatedFiles(entry, dirs, extensions);
+
+        //then
+        assertEquals(Collections.singletonList(Paths.get("src/test/resources/org/jabref/logic/importer/unlinkedFilesTestFolder/directory/subdirectory/2003_Hippel_209.pdf")),
+                result);
+    }
+
+    @Test
+    public void testAuthorWithDiacritics() {
+        //given
+        BibEntry localEntry = new BibEntry(BibtexEntryTypes.ARTICLE.getName());
+        localEntry.setCiteKey("Grazulis2017");
+        localEntry.setField("year", "2017");
+        localEntry.setField("author", "Gražulis, Saulius and O. Kitsune");
+        localEntry.setField("pages", "726--729");
+
+        List<String> extensions = Collections.singletonList("pdf");
+
+        List<Path> dirs = Collections.singletonList(Paths.get(filesDirectory));
+        RegExpBasedFileFinder fileFinder = new RegExpBasedFileFinder("**/[year]_[auth]_[firstpage]\\\\.[extension]", ',');
+
+        //when
+        List<Path> result = fileFinder.findAssociatedFiles(localEntry, dirs, extensions);
+
+        //then
+        assertEquals(Collections.singletonList(Paths.get("src/test/resources/org/jabref/logic/importer/unlinkedFilesTestFolder/directory/subdirectory/2017_Gražulis_726.pdf")),
+                result);
+    }
+
+    @Test
     public void testFindFileInSubdirectory() {
         //given
         BibEntry localEntry = new BibEntry(BibtexEntryTypes.ARTICLE.getName());
@@ -103,33 +141,6 @@ public class RegExpBasedFileFinderTests {
     }
 
     @Test
-    public void testFieldAndFormat() {
-        assertEquals("Eric von Hippel and Georg von Krogh",
-                RegExpBasedFileFinder.getFieldAndFormat("[author]", entry, database, ','));
-
-        assertEquals("Eric von Hippel and Georg von Krogh",
-                RegExpBasedFileFinder.getFieldAndFormat("author", entry, database, ','));
-
-        assertEquals("", RegExpBasedFileFinder.getFieldAndFormat("[unknownkey]", entry, database,
-                ','));
-
-        assertEquals("", RegExpBasedFileFinder.getFieldAndFormat("[:]", entry, database, ','));
-
-        assertEquals("", RegExpBasedFileFinder.getFieldAndFormat("[:lower]", entry, database,
-                ','));
-
-        assertEquals("eric von hippel and georg von krogh",
-                RegExpBasedFileFinder.getFieldAndFormat("[author:lower]", entry, database,
-                        ','));
-
-        assertEquals("HipKro03", RegExpBasedFileFinder.getFieldAndFormat("[bibtexkey]", entry, database,
-                ','));
-
-        assertEquals("HipKro03", RegExpBasedFileFinder.getFieldAndFormat("[bibtexkey:]", entry, database,
-                ','));
-    }
-
-    @Test
     public void testExpandBrackets() {
 
         assertEquals("", RegExpBasedFileFinder.expandBrackets("", entry, database, ','));
@@ -150,6 +161,11 @@ public class RegExpBasedFileFinderTests {
 
         assertEquals(
                 "Eric von Hippel and Georg von Krogh have published Open Source Software and the \"Private-Collective\" Innovation Model: Issues for Organization Science in Organization Science.",
+                RegExpBasedFileFinder.expandBrackets("[author] have published [fulltitle] in [journal].", entry, database,
+                        ','));
+
+        assertEquals(
+                "Eric von Hippel and Georg von Krogh have published Open Source Software and the \"Private Collective\" Innovation Model: Issues for Organization Science in Organization Science.",
                 RegExpBasedFileFinder.expandBrackets("[author] have published [title] in [journal].", entry, database,
                         ','));
     }
